refactor: rename props in state
