+++ src/state/redux/profiles/selectors/notes.ts
